Added threading to scene update, merged scene with boxThreadManager
Threads now work! they work slowly, and this can definitely be iterated upon, but it seems time for a writeup.
To Do: I think thread waiting can use mutexes better, and the whole update seems split - there is definitely work to be done here, but we are in a good state